// providers/socket-types.tsx
export interface AttachmentsSendingProps {
	file: File;
	_id: string;
	name: string;
	preview: string;
	size: string | number;
	type: string;
	progress?: number;
	createdAt?: string;
}